FEATURE: sample-audiovideo-视频图片选择器，播放视频

+++ sample-audiovideo/src/main/java/app/allever/android/learning/audiovideo/StatusListener.kt
@@ -1,7 +1,7 @@
 package app.allever.android.learning.audiovideo
 
 interface StatusListener {
-    fun onPrepare(duration: Int)
+    fun onPrepare(duration: Long)
     fun onVideoPlay()
     fun onVideoPause()
     fun onVideoError()
